chore: rewriting logic to own class, addig tests
