+++ package.json
@@ -48,15 +48,15 @@
     "lit-html": "^1.1.2"
   },
   "devDependencies": {
-    "@advanced-rest-client/arc-models": "^3.0.0-preview.8",
     "@advanced-rest-client/arc-data-export": "^3.0.0",
     "@advanced-rest-client/arc-data-generator": "^3.0.3",
     "@advanced-rest-client/arc-demo-helper": "^1.0.14",
-    "@advanced-rest-client/history-menu": "^3.0.1",
-    "@advanced-rest-client/saved-menu": "^3.0.0",
+    "@advanced-rest-client/arc-models": "^3.0.0-preview.8",
     "@advanced-rest-client/eslint-config": "^1.1.2",
+    "@advanced-rest-client/history-menu": "^3.0.1",
     "@advanced-rest-client/http-method-selector": "^3.0.0",
     "@advanced-rest-client/prettier-config": "^0.1.0",
+    "@advanced-rest-client/saved-menu": "^3.0.0",
     "@advanced-rest-client/testing-karma-sl": "^1.0.4",
     "@anypoint-web-components/anypoint-button": "^1.0.10",
     "@anypoint-web-components/anypoint-checkbox": "^1.0.1",
@@ -69,6 +69,7 @@
     "@open-wc/testing": "^2.3.3",
     "@open-wc/testing-karma": "^3.1.29",
     "@polymer/gen-typescript-declarations": "^1.6.2",
+    "@polymer/iron-test-helpers": "^3.0.1",
     "@polymer/paper-toast": "^3.0.0",
     "@polymer/paper-toggle-button": "^3.0.1",
     "deepmerge": "^4.0.0",
