--- test/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-  <script src="../../../mocha/mocha.js"></script>
-  <script src="../../../wct-mocha/wct-mocha.js"></script>
-</head>
-
-<body>
-  <script>
-  /* global WCT */
-  // Load and run all tests (.html, .js):
-  WCT.loadSuites([
-    'history-panel-test.html',
-    'history-panel-list-test.html',
-    'export-dialog-test.html'
-  ]);
-  </script>
-</body>
-
-</html>
